requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.16.4
python-dotenv==1.1.1
psycopg[binary]==3.2.9
greenlet==3.1.1 